Optimize simulation (#961)

+++ asreview/entry_points/simulate.py
@@ -13,13 +13,12 @@
 # limitations under the License.
 
 
+"""Simulation entry point and utils."""
+
 import json
 import logging
 from datetime import datetime
 from pathlib import Path
-from pathlib import PurePath
-
-import numpy as np
 
 from asreview.compat import convert_id_to_idx
 from asreview.config import ASCII_LOGO
@@ -32,9 +31,6 @@ from asreview.config import DEFAULT_N_PRIOR_INCLUDED
 from asreview.config import DEFAULT_QUERY_STRATEGY
 from asreview.config import EMAIL_ADDRESS
 from asreview.config import GITHUB_PAGE
-from asreview.config import KERAS_MODELS
-from asreview.config import LABEL_NA
-from asreview.data import ASReviewData
 from asreview.data import load_data
 from asreview.entry_points.base import BaseEntryPoint
 from asreview.entry_points.base import _base_parser
@@ -45,10 +41,13 @@ from asreview.models.feature_extraction import get_feature_model
 from asreview.models.query import get_query_model
 from asreview.review.simulate import ReviewSimulate
 from asreview.settings import ASReviewSettings
+from asreview.state import open_state
+from asreview.state.errors import StateNotFoundError
+from asreview.state.paths import get_data_file_path
 from asreview.state.paths import get_data_path
+from asreview.state.paths import get_feature_matrix_path
 from asreview.state.paths import get_project_file_path
 from asreview.state.utils import init_project_folder_structure
-from asreview.state.utils import open_state
 from asreview.types import type_n_queries
 from asreview.utils import get_random_state
 from asreview.webapp.utils.io import read_data
@@ -68,9 +67,10 @@ ASCII_MSG_SIMULATE = """
 """.format(GITHUB_PAGE, EMAIL_ADDRESS)  # noqa
 
 
-def review_finished(project_path, review_id=None):
-    """Mark a review in the project as finished. If no review_id is given,
-    mark the first review as finished.
+def _mark_review_finished(project_path, review_id=None):
+    """Mark a review in the project as finished.
+
+    If no review_id is given, mark the first review as finished.
 
     Arguments
     ---------
@@ -96,6 +96,88 @@ def review_finished(project_path, review_id=None):
         json.dump(project_config, f)
 
 
+def _is_review_finished(project_path, review_id=None):
+    """Check if the given review is finished."""
+    project_path = Path(project_path)
+    with open(get_project_file_path(project_path), 'r') as f:
+        project_config = json.load(f)
+
+    if review_id is None:
+        review_index = 0
+    else:
+        review_index = [x['id']
+                        for x in project_config['reviews']].index(review_id)
+
+    review_info = project_config['reviews'][review_index]
+
+    return ('review_finished' in review_info.keys()) & \
+        review_info['review_finished']
+
+
+def _get_dataset_path_from_args(args_dataset):
+    """Remove 'benchmark:' from the dataset name and add .csv suffix.
+
+    Parameters
+    ----------
+    args_dataset : str
+        Name of the dataset.
+
+    Returns
+    -------
+    str
+        Dataset name without 'benchmark:' if it started with that,
+        and with .csv suffix.
+    """
+    if args_dataset.startswith('benchmark:'):
+        args_dataset = args_dataset[10:]
+
+    return Path(args_dataset).with_suffix('.csv').name
+
+
+def _is_partial_simulation(args):
+    """Check for partial simulation.
+
+    Check if there already is a project file with data of
+    the simulation with given args.
+
+    Arguments
+    ----------
+    args : output of argparse
+
+    Returns
+    -------
+    bool
+        Returns True if there is a project file at args.state_file, with
+        the given dataset, and a review with the given settings, which is not
+        marked as finished.
+    """
+    try:
+        with open_state(args.state_file) as state:
+            settings = state.settings
+    except StateNotFoundError:
+        return False
+
+    # Check if the datasets have the same name.
+    try:
+        project_data_file_name = get_data_file_path(args.state_file).name
+    except Exception:
+        return False
+    args_data_file_name = _get_dataset_path_from_args(args.dataset)
+    if project_data_file_name != args_data_file_name:
+        return False
+
+    # Check if the feature matrix is available.
+    feature_extraction_method = settings.feature_extraction
+    try:
+        feature_matrix_file = get_feature_matrix_path(args.state_file)
+    except FileNotFoundError:
+        return False
+    if not feature_matrix_file.name.startswith(feature_extraction_method):
+        return False
+
+    return True
+
+
 def _set_log_verbosity(verbose):
     if verbose == 0:
         logging.getLogger().setLevel(logging.WARNING)
@@ -106,9 +188,11 @@ def _set_log_verbosity(verbose):
 
 
 class SimulateEntryPoint(BaseEntryPoint):
+    """Entrypoint for simulation."""
+
     description = "Simulate the performance of ASReview."
 
-    def execute(self, argv):
+    def execute(self, argv):  # noqa
 
         # parse arguments
         parser = _simulate_parser()
@@ -125,7 +209,6 @@ class SimulateEntryPoint(BaseEntryPoint):
         # print intro message
         print(ASCII_LOGO + ASCII_MSG_SIMULATE)
 
-        # read dataset into memory
         if args.dataset == "":
 
             with open_state(args.state_file) as state:
@@ -145,58 +228,73 @@ class SimulateEntryPoint(BaseEntryPoint):
             feature_model = get_feature_model(settings.feature_extraction)
 
         else:
+
             as_data = load_data(args.dataset)
 
             if len(as_data) == 0:
                 raise ValueError("Supply at least one dataset"
                                  " with at least one record.")
 
-            # rewrite to ProjectAPI
-            init_project_folder_structure(args.state_file, project_mode='simulate')
+            if not _is_partial_simulation(args):
+                # TODO: refactor these functions into a project module
+                init_project_folder_structure(args.state_file,
+                                              project_mode='simulate')
 
-            # Add the dataset to the project file.
-            as_data.to_csv(
-                Path(
-                    get_data_path(args.state_file),
-                    Path(args.dataset).with_suffix(".csv").name
+                # Add the dataset to the project file.
+                dataset_path = _get_dataset_path_from_args(args.dataset)
+
+                as_data.to_csv(
+                    Path(
+                        get_data_path(args.state_file),
+                        dataset_path
+                    )
                 )
-            )
+                # Update the project.json.
+                project_file_path = get_project_file_path(args.state_file)
+                with open(project_file_path, 'r') as f:
+                    project_json = json.load(f)
+
+                project_json['dataset_path'] = dataset_path
+
+                with open(project_file_path, 'w') as f:
+                    json.dump(project_json, f)
 
             # create a new settings object from arguments
-            settings = ASReviewSettings(model=args.model,
-                                        n_instances=args.n_instances,
-                                        n_queries=args.n_queries,
-                                        n_papers=args.n_papers,
-                                        n_prior_included=args.n_prior_included,
-                                        n_prior_excluded=args.n_prior_excluded,
-                                        query_strategy=args.query_strategy,
-                                        balance_strategy=args.balance_strategy,
-                                        feature_extraction=args.feature_extraction,
-                                        mode="simulate",
-                                        data_fp=None)
+            settings = ASReviewSettings(
+                model=args.model,
+                n_instances=args.n_instances,
+                n_queries=args.n_queries,
+                n_papers=args.n_papers,
+                n_prior_included=args.n_prior_included,
+                n_prior_excluded=args.n_prior_excluded,
+                query_strategy=args.query_strategy,
+                balance_strategy=args.balance_strategy,
+                feature_extraction=args.feature_extraction,
+                mode="simulate",
+                data_fp=None)
             settings.from_file(args.config_file)
 
             # Initialize models.
             random_state = get_random_state(args.seed)
             classifier_model = get_classifier(settings.model,
-                                              **settings.model_param,
-                                              random_state=random_state)
+                                              random_state=random_state,
+                                              **settings.model_param)
             query_model = get_query_model(settings.query_strategy,
-                                          **settings.query_param,
-                                          random_state=random_state)
+                                          random_state=random_state,
+                                          **settings.query_param)
             balance_model = get_balance_model(settings.balance_strategy,
-                                              **settings.balance_param,
-                                              random_state=random_state)
+                                              random_state=random_state,
+                                              **settings.balance_param)
             feature_model = get_feature_model(settings.feature_extraction,
-                                              **settings.feature_param,
-                                              random_state=random_state)
+                                              random_state=random_state,
+                                              **settings.feature_param)
 
             # TODO{Deprecate and integrate with the model}
             # LSTM models need embedding matrices.
             if classifier_model.name.startswith("lstm-"):
                 texts = as_data.texts
-                classifier_model.embedding_matrix = feature_model.get_embedding_matrix(
-                    texts, args.embedding_fp)
+                classifier_model.embedding_matrix = feature_model.\
+                    get_embedding_matrix(texts, args.embedding_fp)
 
             # prior knowledge
             if args.prior_idx is not None and args.prior_record_id is not None and \
@@ -213,26 +311,27 @@ class SimulateEntryPoint(BaseEntryPoint):
                 preview = preview_record(as_data.record(prior_record_id))
                 print(f"{prior_record_id} - {preview}")
 
-        # Initialize the review class.
-        reviewer = ReviewSimulate(as_data,
-                                  state_file=args.state_file,
-                                  model=classifier_model,
-                                  query_model=query_model,
-                                  balance_model=balance_model,
-                                  feature_model=feature_model,
-                                  n_papers=args.n_papers,
-                                  n_instances=args.n_instances,
-                                  n_queries=args.n_queries,
-                                  prior_indices=prior_idx,
-                                  n_prior_included=args.n_prior_included,
-                                  n_prior_excluded=args.n_prior_excluded,
-                                  init_seed=args.init_seed)
+            # Initialize the review class.
+            reviewer = ReviewSimulate(as_data,
+                                      state_file=args.state_file,
+                                      model=classifier_model,
+                                      query_model=query_model,
+                                      balance_model=balance_model,
+                                      feature_model=feature_model,
+                                      n_papers=args.n_papers,
+                                      n_instances=args.n_instances,
+                                      n_queries=args.n_queries,
+                                      prior_indices=prior_idx,
+                                      n_prior_included=args.n_prior_included,
+                                      n_prior_excluded=args.n_prior_excluded,
+                                      init_seed=args.init_seed,
+                                      write_interval=args.write_interval)
 
-        # Start the review process.
-        reviewer.review()
+            # Start the review process.
+            reviewer.review()
 
-        # Mark review as finished.
-        review_finished(args.state_file)
+            # Mark review as finished.
+            _mark_review_finished(args.state_file)
 
 
 DESCRIPTION_SIMULATE = """
@@ -355,6 +454,14 @@ def _simulate_parser(prog="simulate", description=DESCRIPTION_SIMULATE):
         default=0,
         type=int,
         help="Verbosity"
+    )
+    parser.add_argument(
+        "--write_interval", "-w",
+        default=None,
+        type=int,
+        help="The simulation data will be written away after each set of this"
+             "many labeled records. By default only writes away data at the end"
+             "of the simulation to make it as fast as possible."
     )
 
     return parser
